+++ flutter_inappwebview_platform_interface/lib/src/in_app_webview/platform_inappwebview_widget.dart
@@ -166,6 +166,7 @@ class PlatformInAppWebViewWidgetCreationParams
   ///- Android native WebView
   ///- iOS
   ///- Web
+  ///- MacOS
   final PlatformHeadlessInAppWebView? headlessWebView;
 
   ///Used to keep alive this WebView.
@@ -175,6 +176,8 @@ class PlatformInAppWebViewWidgetCreationParams
   ///**Officially Supported Platforms/Implementations**:
   ///- Android native WebView
   ///- iOS
+  ///- MacOS
+  ///- Windows
   final InAppWebViewKeepAlive? keepAlive;
 
   ///Used to prevent gesture delay on iOS caused by Flutter's gestures handling