+++ flutter_inappwebview_windows/windows/custom_platform_view/custom_platform_view.cc
@@ -116,7 +116,7 @@ namespace flutter_inappwebview_plugin
     flutter::TextureRegistrar* texture_registrar,
     GraphicsContext* graphics_context,
     HWND hwnd,
-    std::unique_ptr<flutter_inappwebview_plugin::InAppWebView> webView)
+    std::shared_ptr<flutter_inappwebview_plugin::InAppWebView> webView)
     : hwnd_(hwnd), view(std::move(webView)), texture_registrar_(texture_registrar)
   {
 #ifdef HAVE_FLUTTER_D3D_TEXTURE
@@ -180,7 +180,6 @@ namespace flutter_inappwebview_plugin
         },
         [this](const flutter::EncodableValue* arguments)
         {
-          event_sink_ = nullptr;
           return nullptr;
         });
 
@@ -191,6 +190,7 @@ namespace flutter_inappwebview_plugin
   {
     debugLog("dealloc CustomPlatformView");
     method_channel_->SetMethodCallHandler(nullptr);
+    event_sink_ = nullptr;
     texture_registrar_->UnregisterTexture(texture_id_);
   }
 
@@ -209,8 +209,10 @@ namespace flutter_inappwebview_plugin
       {
         const auto& name = GetCursorName(cursor);
         const auto event = flutter::EncodableValue(
-          flutter::EncodableMap { {flutter::EncodableValue(kEventType),
-          flutter::EncodableValue("cursorChanged")},
+          flutter::EncodableMap { {
+              flutter::EncodableValue(kEventType),
+                flutter::EncodableValue("cursorChanged")
+            },
           { flutter::EncodableValue(kEventValue), name }});
         EmitEvent(event);
       });