+++ flutter_inappwebview_windows/windows/custom_platform_view/custom_platform_view.h
@@ -17,13 +17,13 @@ namespace flutter_inappwebview_plugin
   public:
     static inline const wchar_t* CLASS_NAME = L"CustomPlatformView";
 
-    const std::unique_ptr<flutter_inappwebview_plugin::InAppWebView> view;
+    std::shared_ptr<flutter_inappwebview_plugin::InAppWebView> view;
 
     CustomPlatformView(flutter::BinaryMessenger* messenger,
       flutter::TextureRegistrar* texture_registrar,
       GraphicsContext* graphics_context,
       HWND hwnd,
-      std::unique_ptr<flutter_inappwebview_plugin::InAppWebView> webView);
+      std::shared_ptr<flutter_inappwebview_plugin::InAppWebView> webView);
     ~CustomPlatformView();
 
     TextureBridge* texture_bridge() const { return texture_bridge_.get(); }