+++ flutter_inappwebview_windows/windows/in_app_webview/webview_channel_delegate.h
@@ -6,6 +6,7 @@
 
 #include "../types/base_callback_result.h"
 #include "../types/channel_delegate.h"
+#include "../types/create_window_action.h"
 #include "../types/navigation_action.h"
 #include "../types/web_resource_error.h"
 #include "../types/web_resource_request.h"
@@ -34,6 +35,12 @@ namespace flutter_inappwebview_plugin
       ~CallJsHandlerCallback() = default;
     };
 
+    class CreateWindowCallback : public BaseCallbackResult<const bool> {
+    public:
+      CreateWindowCallback();
+      ~CreateWindowCallback() = default;
+    };
+
     WebViewChannelDelegate(InAppWebView* webView, flutter::BinaryMessenger* messenger);
     WebViewChannelDelegate(InAppWebView* webView, flutter::BinaryMessenger* messenger, const std::string& name);
     ~WebViewChannelDelegate();
@@ -44,6 +51,7 @@ namespace flutter_inappwebview_plugin
 
     void onLoadStart(const std::optional<std::string>& url) const;
     void onLoadStop(const std::optional<std::string>& url) const;
+    void onProgressChanged(const int64_t& progress) const;
     void shouldOverrideUrlLoading(std::shared_ptr<NavigationAction> navigationAction, std::unique_ptr<ShouldOverrideUrlLoadingCallback> callback) const;
     void onReceivedError(std::shared_ptr<WebResourceRequest> request, std::shared_ptr<WebResourceError> error) const;
     void onReceivedHttpError(std::shared_ptr<WebResourceRequest> request, std::shared_ptr<WebResourceResponse> error) const;
@@ -52,6 +60,8 @@ namespace flutter_inappwebview_plugin
     void onCallJsHandler(const std::string& handlerName, const std::string& args, std::unique_ptr<CallJsHandlerCallback> callback) const;
     void onConsoleMessage(const std::string& message, const int64_t& messageLevel) const;
     void onDevToolsProtocolEventReceived(const std::string& eventName, const std::string& data) const;
+    void onCreateWindow(std::shared_ptr<CreateWindowAction> createWindowAction, std::unique_ptr<CreateWindowCallback> callback) const;
+    void onCloseWindow() const;
   };
 }