+++ flutter_inappwebview_windows/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.0
+
+- Added support for keeping alive InAppWebView widgets
+- Added onProgressChanged, onCreateWindow, onCloseWindow support
+
 ## 0.1.0
 
 - Initial release